backend/requirements.txt
backboard-sdk==1.5.0
beautifulsoup4==4.12.3
fastapi==0.129.0
httpx==0.28.1
ipykernel==6.29.5
pillow==12.1.1
pydantic==2.12.5
pytesseract==0.3.13
python-dotenv==1.2.1
requests==2.32.5
uvicorn==0.40.0